Moving default map extensions into Mapping namespace
