+++ src/Gate/Mapping/Delegates.cs
