+++ src/Gate/Mapping/UrlMapper.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Gate.Owin;
 
-namespace Gate.Builder
+namespace Gate.Mapping
 {
     public class UrlMapper
     {
